test: kinematic physics for player

+++ src/tiled/map.cpp
@@ -132,7 +132,7 @@ void Map::register_tileset(const char* filename, const char* tileset_string) {
 
     std::vector<Tile::Data> tiles;
     for (size_t i = 0; i < tile_count; ++i) {
-        int x = i % column_count, y = i / column_count;
+        float x = i % column_count, y = i / column_count;
         tiles.emplace_back(Render::Rect{x * tile_width, y * tile_height, tile_width, tile_height},
                            Map::get_texture(image_file));
     }
